Got rid of ViewModelCommand and switched entirely over to Command.
