+++ cpvc/UI/BookmarksViewModel.cs
@@ -19,10 +19,10 @@ namespace CPvC.UI
         private HistoryViewItem _selectedItem;
         private Display _display;
 
-        private ViewModelCommand _deleteBookmarksCommand;
-        private ViewModelCommand _deleteBranchesCommand;
-        private ViewModelCommand _jumpToBookmarkCommand;
-        private ViewModelCommand _replayTimelineCommand;
+        private Command _deleteBookmarksCommand;
+        private Command _deleteBranchesCommand;
+        private Command _jumpToBookmarkCommand;
+        private Command _replayTimelineCommand;
 
         public ICommand DeleteBookmarksCommand
         {
@@ -62,28 +62,24 @@ namespace CPvC.UI
             // The initial selected item is set to the current event.
             SelectedItem = Items.FirstOrDefault(i => i.HistoryEvent == _machine.CurrentEvent);
 
-            _replayTimelineCommand = new ViewModelCommand(
+            _replayTimelineCommand = new Command(
                 p => ReplayTimeline(itemSelected),
-                p => (SelectedItem as HistoryViewItem)?.HistoryEvent != null,
-                this, "SelectedItem", null
+                p => (SelectedItem as HistoryViewItem)?.HistoryEvent != null
             );
 
-            _jumpToBookmarkCommand = new ViewModelCommand(
+            _jumpToBookmarkCommand = new Command(
                 p => JumpToBookmark(itemSelected),
-                p => (SelectedItem as HistoryViewItem)?.HistoryEvent.Bookmark != null,
-                this, "SelectedItem", null
+                p => (SelectedItem as HistoryViewItem)?.HistoryEvent.Bookmark != null
             );
 
-            _deleteBookmarksCommand = new ViewModelCommand(
+            _deleteBookmarksCommand = new Command(
                 p => DeleteBookmarks(),
-                p => (SelectedItem as HistoryViewItem)?.HistoryEvent.Bookmark != null,
-                this, "SelectedItem", null
+                p => (SelectedItem as HistoryViewItem)?.HistoryEvent.Bookmark != null
             );
 
-            _deleteBranchesCommand = new ViewModelCommand(
+            _deleteBranchesCommand = new Command(
                 p => DeleteBranches(),
-                p => (SelectedItem as HistoryViewItem)?.HistoryEvent != null,
-                this, "SelectedItem", null
+                p => (SelectedItem as HistoryViewItem)?.HistoryEvent != null
             );
         }
 
@@ -309,7 +305,14 @@ namespace CPvC.UI
 
         protected void OnPropertyChanged(string name)
         {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
+            PropertyChangedEventArgs args = new PropertyChangedEventArgs(name);
+
+            _deleteBookmarksCommand?.InvokeCanExecuteChanged(this, args);
+            _deleteBranchesCommand?.InvokeCanExecuteChanged(this, args);
+            _jumpToBookmarkCommand?.InvokeCanExecuteChanged(this, args);
+            _replayTimelineCommand?.InvokeCanExecuteChanged(this, args);
+
+            PropertyChanged?.Invoke(this, args);
         }
     }
 }
